simplify server slice (rm ws mgmt - that happens in the ws service)
