+++ app-design-v2/src/store/slices/server/index.ts
@@ -1,4 +1,4 @@
-export * from './connection-slice.ts';
+export * from './server-slice.ts';
 export * from './server-types';
 export * from './server-selectors';
 export * from './server-thunks';
