+++ app-design-v2/src/store/store.ts
@@ -2,7 +2,7 @@ import {configureStore} from "@reduxjs/toolkit";
 import {framerateSlice} from "./slices/framerate/framerate-slice";
 import {cameraSlice} from "./slices/cameras/cameras-slice";
 import {recordingSlice} from "./slices/recording/recording-slice";
-import {connectionSlice} from "@/store/slices/connection/connection-slice.ts";
+import {serverSlice} from "@/store/slices/server/server-slice.ts";
 import {logRecordsSlice} from "./slices/log-records/log-records-slice";
 
 export const store = configureStore({
@@ -11,7 +11,7 @@ export const store = configureStore({
         recording: recordingSlice.reducer,
         framerate: framerateSlice.reducer,
         logs: logRecordsSlice.reducer,
-        server: connectionSlice.reducer,
+        server: serverSlice.reducer,
     }
 });
